requirements.txt
alembic==1.8.0
attrs==21.4.0
black==22.10.0
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.0
greenlet==1.1.2
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
